Release v0.2.0 (#43)
* Added `acc` and `make_acc_group` fixtures
([#42](https://github.com/databrickslabs/pytester/issues/42)). In this
release, we have added two new fixtures, `acc` and `make_acc_group`, to
the open-source library. The `acc` fixture provides a Databricks
AccountClient object for use in tests, which can interact with the
Databricks account API and automatically determines the account host
from the `DATABRICKS_HOST` environment variable. The `make_acc_group`
fixture is used for managing Databricks account groups, creating them
with specified members and roles, and automatically deleting them after
the test is complete. This fixture mirrors the behavior of the
`make_group` fixture but interacts with the account client instead of
the workspace client. These fixtures enable more comprehensive
integration tests for the `acc` object and its various methods,
enhancing the testing and management of Databricks account groups.