Release v0.2.2 (#49)
* Use watchdog timeout to catalog properties
([#48](https://github.com/databrickslabs/pytester/issues/48)). This pull
request introduces a new `RemoveAfter` property for catalogs, which
allows for marking them for skipping by the watchdog. This change
addresses the current implementation gap, which does not explicitly
indicate when catalogs are being used. The new property will specify the
time from which objects can be purged. A corresponding fixture
`watchdog_remove_after` has been added to the list of available
fixtures, and the `make_catalog` fixture has been updated to include
this new property. Additionally, a timeout mechanism for catalogs has
been implemented, which improves the system's efficiency and safety by
marking catalogs as in use. A test for the `make_catalog` function has
been included to ensure that the `RemoveAfter` entry is correctly added
to the catalog properties. However, the specific call parameters for the
`catalogs.create` method cannot be accurately determined in the test.
* use tags instead of name suffix for queries
([#47](https://github.com/databrickslabs/pytester/issues/47)). This
release introduces updates to the testing library for Databricks,
enhancing the naming conventions for queries to improve readability and
comprehension. The previous implementation used name suffixes, which
have been replaced with watchdog query tags. The `watchdog_purge_suffix`
fixture has been renamed to `watchdog_remove_after`, and the new
`make_query` fixture has been added to the documentation. In addition,
the `make_query` and `create` functions now accept an optional `tags`
argument, and the query name is generated with a unique identifier. If
`tags` are provided, the `RemoveAfter` tag is added. The
`original_query_tag` is no longer hardcoded in the `create` function and
has been removed. These changes improve the overall user experience and
maintainability of the project.