Release v0.2.3 (#53)
* Support providing name in `make_catalog` fixture
([#52](https://github.com/databrickslabs/pytester/issues/52)). The
`make_catalog` fixture in our open-source library has been updated to
allow users to specify a name for the catalog using a new `name`
parameter. Previously, the catalog was given a random name, but now
users can have more control and customization over catalog names in
their tests. This change includes updates to the docstring and the
addition of unit tests to ensure the fixture behaves as expected with
the new parameter. Additionally, the underlying `call_stateful` function
was updated to expect a callable that returns a generator of callables,
enabling the support for providing a name. The
`test_make_catalog_creates_catalog_with_name` and `test_make_catalog`
tests have been added to verify the behavior of the fixture with the new
`name` parameter.