Improve `make_group`/`make_acc_group` fixture consistency (#50)
## Changes

This PR implements some fixture improvements that were originally
implemented downstream:

- When creating an account or workspace group, we now require the group
to be visible via two subsequent `.get()` and `.list()` calls. This
double-check approach mitigates (but doesn't completely eliminate)
problems with consistency of the APIs for working with groups.
- The `wait_for_provisioning` argument to the `make_group` and
`make_acc_group` has now been removed: we always wait. (For
compatibility the argument is still accepted but will trigger a
deprecation warning.)

An incidental change is the internal unit-test plumbing can now be
provided with mock fixtures specific to the test; this is needed to
ensure the double-check implementation can be tested from the unit
tests.

### Linked issues

Supersedes databrickslabs/ucx#2634.

### Tests

- added/updated unit tests
- existing integration tests