Release v0.3.1 (#67)
* Updated databrickslabs/sandbox requirement to acceptance/v0.3.1
([#64](https://github.com/databrickslabs/pytester/issues/64)). In this
pull request, we are updating the requirement for the
`databrickslabs/sandbox` package to version `acceptance/v0.3.1`. This
update is necessary to resolve any conflicts and ensure compatibility
with the latest version of the package. The update includes several bug
fixes, dependency updates, and the addition of install instructions in
the changelog. The package also includes new git-related libraries and
modifications to the README to explain how to use the package with the
`databricks labs sandbox` command. Additionally, there are dependency
updates for `golang.org/x/crypto` in the `go-libs` and
`runtime-packages` directories. Keeping dependencies up-to-date is
important for taking advantage of new features and bug fixes, so it is
recommended to merge this pull request and address any conflicts or
issues that may arise. This update is generated by Dependabot, a tool
used to keep dependencies up-to-date, and it will handle any conflicts
as long as the pull request is not modified. The
`databrickslabs/sandbox` package is used for building and testing the
project.
* Updating `make_schema` fixture to include location to create managed
schema ([#66](https://github.com/databrickslabs/pytester/issues/66)). In
this update, the `make_schema` fixture has been enhanced to include an
optional `location` parameter, enabling users to specify the location
for creating a managed schema. This change addresses issue
[#6](https://github.com/databrickslabs/pytester/issues/6)
* [chore] update acceptance.yml and remove circular downstreads. In this
release, the `acceptance.yml` file in the `.github/workflows` directory
has been updated to use the `acceptance/v0.3.1` Docker image for the
`Run integration tests` job, replacing the previous `acceptance/v0.2.2`
version. This change ensures that the latest version of the integration
tests are being run, which may include bug fixes, new features, and
other improvements. The `timeout` parameter has been removed from the
job, which may result in longer test execution times but ensures that
the tests have enough time to complete. Additionally, the
`ARM_CLIENT_ID` and `GITHUB_TOKEN` secrets are now passed as environment
variables to the job. The `downstreams.yml` file has also been updated
to remove circular dependencies and streamline the build process, while
maintaining the use of the `ubuntu-latest` runtime and `fail-fast`
option. The `Run nightly tests` job has been updated to use the
`acceptance/v0.3.1` Docker image, which may improve test execution
during the nightly build due to potential bug fixes, new functionality,
or better handling of timeout scenarios. The `create_issues` parameter
remains set to true, ensuring that any test failures will result in the
creation of GitHub issues to track them. These changes aim to improve
the reliability and efficiency of the build and test processes.

Dependency updates:

* Updated databrickslabs/sandbox requirement to acceptance/v0.3.1
([#64](https://github.com/databrickslabs/pytester/pull/64)).