Compatibility fix for Databricks SDK 0.51 (#158)
## Changes

[Databricks SDK
0.51](https://github.com/databricks/databricks-sdk-py/releases/tag/v0.51.0)
included some breaking changes, including removal of an enumeration
where a set of well-known string constants must now be used instead:
> Breaking] Removed large, medium and small enum values for
databricks.sdk.service.serving.ServedModelInputWorkloadSize.

The change is not forward-compatible, so in addition to updating the
fixture we need to bump the minimum-required SDK version because older
versions expect the enum.

### Credits

This was earlier discovered and reported by @cornzyblack as #156.

### Linked issues

Resolves #156.
Supersedes #157.

This is also causing CI test failures on downstream projects including:

 - [databrickslabs/lsql](https://github.com/databrickslabs/lsql)
-
[databrickslabs/blueprint](https://github.com/databrickslabs/blueprint)

### Tests

- existing unit tests