Release v0.7.4 (#215)
* Increased the number of random characters used when creating resources
to minimize collisions
([#214](https://github.com/databrickslabs/pytester/issues/214)). The
continuous integration workflow has been updated to install a newer
version of the hatch tool, specifically version 1.15.0, replacing the
previous version 1.9.4, to avoid compatibility issues with click.
Additionally, the code has been modified to increase the number of
random characters used when creating resources from 4 to 8, aiming to
minimize collisions, particularly in larger projects that rely heavily
on PyTester. This change affects various resource creation methods and
is intended to reduce the likelihood of name collisions and improve the
overall stability of the system. The changes are covered by regression
testing and have been manually tested to ensure their correctness and
effectiveness.